Extract pipeline definition helpers and creation from the renderer
